Update instructions.html
I deleted the sentence that the points will be accumulated because I felt that it was a bit ambiguous, as if the points of A and B would be added together. I also think we do not need the sentence.

Importantly, I took out the sentence that the goal of the experiment is to maximize points. I feel this sentence belongs in the meta prompt. 

In the meta prompt, we could say something like:
"The goal of the experiment is to maximize your points. However, you will be interacting with other participants who will also try to maximize their points. When making decisions participants can also take into account how these decisions will affect other participants" Or something like that.



If you think this is mistake, feel free to add it back.